[FIXING] Add the YoonLine Function and Fixed Error

+++ YoonFactory/YoonCommon/YoonRect.cs
@@ -215,7 +215,7 @@ namespace YoonFactory
             return false;
         }
 
-        public int Area()
+        public double Area()
         {
             return Width * Height;
         }
